feat: migrate /childAlert endpoint to AlertInfoController
- Migrated getChildrenByAddress method from PersonController to AlertInfoController
- Moved related business logic from PersonService to AlertInfoService
- Enhanced getPersonByAddress in InMemoryPersonRepository with trim() and replaceAll("\\s+", " ") for better input handling
- Applied same normalization in getMedicalRecordByFirstNameAndLastName (InMemoryMedicalRecordRepository) and findByLastName (InMemoryPersonRepository)
- Verified endpoint functionality at /childAlert?address=<address> via Postman
